Fix multi-line comments in JSON-format translation files
The `note` field in the JSON translation files can include newlines which need
to handled when converting to PO, to ensure that each new line of a comment
begins with a comment identifier. When converting from PO to JSON, multi-line
comments need to be joined together into a single string separated by newlines.